S3 - Refactor tests
